Refactoring to match habits
